Move Scope and Aliases to separate files
